updates to the invite key verification and operations for negative cases; new test cases that cover an invite without a key setup, an invite with a key setup but wrong temporary key; update to the successful invite unit user test case for checking that the keys of unit projects are shared to the invite

+++ dds_web/api/schemas/user_schemas.py
@@ -196,27 +196,33 @@ class NewUserSchema(marshmallow.Schema):
 
         # Verify and transfer invite keys to the new user
         new_user.temporary_key = dds_web.security.auth.verify_invite_key(token)
-        new_user.nonce = invite.nonce
-        new_user.public_key = invite.public_key
-        new_user.private_key = invite.private_key
-        for project_invite_key in invite.project_invite_keys:
-            project_user_key = models.ProjectUserKeys(
-                project_id=project_invite_key.project_id,
-                user_id=new_user.username,
-                key=project_invite_key.key,
-            )
-            db.session.add(project_user_key)
-            db.session.delete(project_invite_key)
+        if new_user.temporary_key:
+            new_user.nonce = invite.nonce
+            new_user.public_key = invite.public_key
+            new_user.private_key = invite.private_key
+            for project_invite_key in invite.project_invite_keys:
+                project_user_key = models.ProjectUserKeys(
+                    project_id=project_invite_key.project_id,
+                    user_id=new_user.username,
+                    key=project_invite_key.key,
+                )
+                db.session.add(project_user_key)
+                db.session.delete(project_invite_key)
 
-        # TODO decrypt the user private key using the temp key,
-        #  derive a key from the password and encrypt the user private key with the derived key
+            # TODO decrypt the user private key using the temp key,
+            #  derive a key from the password and encrypt the user private key with the derived key
+
+            flask.session.pop("invite_token", None)
+
+            # Delete old invite
+            db.session.delete(invite)
+
+            # Save and return
+            db.session.commit()
+
+            return new_user
 
         flask.session.pop("invite_token", None)
-
-        # Delete old invite
-        db.session.delete(invite)
-
-        # Save and return
+        db.session.delete(new_user)
         db.session.commit()
-
-        return new_user
+        raise ddserr.InviteError(message="Invite key verification has failed!")
